Revert "move metrics_tests too"
This reverts commit 0b90f8a8f4f4495cfc3879a4aeca720b83f8b27c.
